T48: extract SettingsWindowUI from ParsekUI
Move settings window (353 lines, 8 fields, 10 methods) into its own
SettingsWindowUI class in UI/. Settings calls back to ParsekUI for wipe
confirmations and test runner toggle. Loop/unit conversion helpers are
already internal static on ParsekUI — called directly from the new class.
